requirements.txt
amqp==5.1.1
asgiref==3.7.2
auditlog==0.0.5
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-auditlog==2.3.0
django-celery-beat==2.5.0
django-extensions==3.2.3
django-import-export==3.3.1
django-jsonfield==1.4.1
django-moneyfield==0.2.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
docopt==0.6.2
et-xmlfile==1.1.0
idna==3.4
kombu==5.3.2
MarkupPy==1.14
mysqlclient==2.2.0
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.1
pipreqs==0.4.13
prompt-toolkit==3.0.39
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.0.5
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.8
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
